backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
chromadb==0.4.18
google-generativeai==0.3.2
pdfplumber==0.10.3
python-docx==1.1.0
requests==2.31.0
beautifulsoup4==4.12.2
youtube-transcript-api==0.6.2
python-multipart==0.0.6
aiofiles==23.2.1
pydantic-core==2.14.1
